Small bug fix and changes to siege banner
I ran a small test on my computer. Players not in a faction will not be kill after 5 minutes on another factions land because it failed to check whether player's faction was blank. It check if the blank faction was an enemy which would return false and fail to kill the player.

I made a change in the can_claim_parcel function. You could only claim over faction land if your faction is at war with them.

Small bug fix with siege banner. The siege stage started at zero which made it take longer. before the fix it would stay on stage 1 longer then the other stages.